requirements.txt
streamlit==1.26.0
pymongo[srv]==4.5.0
python-dotenv==1.0.0
